Actors - Improving documentation; other improvements

+++ lib/concurrent/actor/behaviour/linking.rb
@@ -1,14 +1,49 @@
 module Concurrent
   module Actor
     module Behaviour
+      # TODO track what is linked, clean when :terminated
+      #   send :linked/:unlinked messages back to build the array of linked actors
 
       # Links the actor to other actors and sends actor's events to them,
-      # like: `:terminated`, `:paused`, errors, etc
-      # TODO example
+      # like: `:terminated`, `:paused`, `:resumed`, errors, etc.
+      # Linked actor needs to handle those messages.
+      #
+      #     listener = AdHoc.spawn name: :listener do
+      #       lambda do |message|
+      #         case message
+      #         when Reference
+      #           if message.ask!(:linked?)
+      #             message << :unlink
+      #           else
+      #             message << :link
+      #           end
+      #         else
+      #           puts "got event #{message.inspect} from #{envelope.sender}"
+      #         end
+      #       end
+      #     end
+      #
+      #     an_actor = AdHoc.spawn name: :an_actor, supervise: true, behaviour_definition: Behaviour.restarting_behaviour_definition do
+      #       lambda { |message| raise 'failed'}
+      #     end
+      #
+      #     # link the actor
+      #     listener.ask(an_actor).wait
+      #     an_actor.ask(:fail).wait
+      #     # unlink the actor
+      #     listener.ask(an_actor).wait
+      #     an_actor.ask(:fail).wait
+      #     an_actor << :terminate!
+      #
+      # produces only two events, other events happened after unlinking
+      #
+      #     got event #<RuntimeError: failed> from #<Concurrent::Actor::Reference /an_actor (Concurrent::Actor::Utils::AdHoc)>
+      #     got event :reset from #<Concurrent::Actor::Reference /an_actor (Concurrent::Actor::Utils::AdHoc)>
       class Linking < Abstract
-        def initialize(core, subsequent)
-          super core, subsequent
+        def initialize(core, subsequent, core_options)
+          super core, subsequent, core_options
           @linked = Set.new
+          @linked.add Actor.current if core_options[:link] != false
         end
 
         def on_envelope(envelope)
@@ -19,6 +54,8 @@ module Concurrent
             unlink envelope.sender
           when :linked?
             @linked.include? envelope.sender
+          when :linked
+            @linked.to_a
           else
             pass envelope
           end
@@ -34,10 +71,11 @@ module Concurrent
           true
         end
 
-        def on_event(event)
-          @linked.each { |a| a << event }
-          @linked.clear if event == :terminated
-          super event
+        def on_event(public, event)
+          event_name, _ = event
+          @linked.each { |a| a << event } if public
+          @linked.clear if event_name == :terminated
+          super public, event
         end
       end
     end
